requirements.txt
wandb
torchmetrics[image]
numpy
pandas
matplotlib
tensorflow
python-dotenv
pillow